% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyvalidinp.R
\name{validinp_character}
\alias{validinp_character}
\title{validate character input}
\usage{
validinp_character(x, pattern = "^[[:alnum:]. _-]*$", many = FALSE)
}
\arguments{
\item{x}{input}

\item{pattern}{that input has to match (regexp)}

\item{many}{TRUE if more than 1 string (checkboxGroupInput)}
}
\value{
validated input
}
\description{
validate character: textInput, selectInput, checkboxGroupInput, radioButtons
}
\examples{
\dontrun{validinp_character(input$txt)}
\dontrun{validinp_character(input$radiobox, pattern="^((ab)|(cd))$")}
\dontrun{validinp_character(input$chkboxgrp, many=TRUE}
}

