% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyvalidinp.R
\name{validinp_Date}
\alias{validinp_Date}
\title{validate Date input}
\usage{
validinp_Date(x, min = -Inf, max = Inf, many = FALSE)
}
\arguments{
\item{x}{input}

\item{min}{min date}

\item{max}{max date}

\item{many}{TRUE if date range (i.e. 2 values)}
}
\value{
validated input
}
\description{
validate Date: dateInput, dateRangeInput
}
\examples{
\dontrun{validinp_Date(input$dt)}
\dontrun{validinp_Date(input$dt_range, many=TRUE)}
}

