% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyvalidinp.R
\name{validinp_logical}
\alias{validinp_logical}
\title{validate logical input}
\usage{
validinp_logical(x)
}
\arguments{
\item{x}{input}
}
\value{
validated input
}
\description{
validate TRUE/FALSE: checkboxInput
}
\examples{
\dontrun{validinp_logical(input$chkbox)}
}

