% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyvalidinp.R
\name{validinp_numeric}
\alias{validinp_numeric}
\title{validate numeric input}
\usage{
validinp_numeric(x, min = -Inf, max = Inf)
}
\arguments{
\item{x}{input}

\item{min}{min value}

\item{max}{max value}
}
\value{
validated input
}
\description{
validate numeric: numericInput, sliderInput, actionButton
}
\examples{
\dontrun{validinp_numeric(input$n)}
\dontrun{validinp_numeric(input$n, min=0, max=10)}
}

